show collectables in owner activity

+++ app/src/main/java/com/qrhunter/PlayerDatabse.java
@@ -180,11 +180,15 @@ public class PlayerDatabse {
     }
 
     public void deleteUser(String username) {
-        players.remove(username);
-        database.collection("Players")
-                .document(username)
-                .delete()
-                .addOnFailureListener(e -> {throw new RuntimeException("Network Error.");});
+        if(players.containsKey(username)) {
+            players.remove(username);
+            database.collection("Players")
+                    .document(username)
+                    .delete()
+                    .addOnFailureListener(e -> {
+                        throw new RuntimeException("Network Error.");
+                    });
+        }
     }
 
 }
